+++ src/walletSdk/Anchor/index.ts
@@ -12,16 +12,16 @@ import {
   AssetNotSupportedError,
 } from "../Exceptions";
 import { Interactive } from "../Interactive";
-import { TomlInfo, parseToml } from "../Toml";
 import { 
   AnchorServiceInfo, 
   AnchorTransaction, 
   GetTransactionParams, 
   GetTransactionsParams, 
+  TomlInfo, 
   TransactionStatus 
 } from "../Types";
 import { Watcher } from "../Watcher";
-import { camelToSnakeCaseObject } from "../Util/camelToSnakeCase";
+import { camelToSnakeCaseObject, parseToml } from "../Utils";
 
 
 // Let's keep this constructor type private as
