+++ src/walletSdk/Types/index.ts
@@ -2,3 +2,4 @@ export * from "./anchor";
 export * from "./auth";
 export * from "./horizon";
 export * from "./interactive";
+export * from "./utils";
