+++ src/walletSdk/Types/utils.ts
@@ -1,0 +1,77 @@
+export type TomlInfo = {
+  version: string;
+  networkPassphrase: string;
+  federationServer: string;
+  authServer: string;
+  transferServer: string;
+  transferServerSep24: string;
+  kycServer: string;
+  webAuthEndpoint: string;
+  signingKey: string;
+  horizonUrl: string;
+  accounts: Array<string>;
+  uriRequestSigningKey: string;
+  directPaymentServer: string;
+  anchorQuoteServer: string;
+  documentation: {
+    orgName: string;
+    orgDba: string;
+    orgUrl: string;
+    orgLogo: string;
+    orgDescription: string;
+    orgPhysicalAddress: string;
+    orgPhysicalAddressAttestation: string;
+    orgPhoneNumber: string;
+    orgPhoneNumberAttestation: string;
+    orgKeybase: string;
+    orgTwitter: string;
+    orgGithub: string;
+    orgOfficialEmail: string;
+    orgSupportEmail: string;
+    orgLicensingAuthority: string;
+    orgLicenseType: string;
+    orgLicenseNumber: string;
+  };
+  principals: Array<{
+    name: string;
+    email: string;
+    keybase: string;
+    telegram: string;
+    twitter: string;
+    github: string;
+    idPhotoHash: string;
+    verificationPhotoHash: string;
+  }>;
+  currencies: Array<{
+    code: string;
+    codeTemplate: string;
+    issuer: string;
+    status: string;
+    displayDecimals: number;
+    name: string;
+    desc: string;
+    conditions: string;
+    image: string;
+    fixedNumber: number;
+    maxNumber: number;
+    isUnlimited: boolean;
+    isAssetAnchored: boolean;
+    anchorAssetType: string;
+    anchorAsset: string;
+    attestationOfReserve: string;
+    redemptionInstructions: string;
+    collateralAddresses: Array<string>;
+    collateralAddressMessages: Array<string>;
+    collateralAddressSignatures: Array<string>;
+    regulated: boolean;
+    approvalServer: string;
+    approvalCriteria: string;
+  }>;
+  validators: Array<{
+    alias: string;
+    displayName: string;
+    publicKey: string;
+    host: string;
+    history: string;
+  }>;
+};
