+++ src/walletSdk/Utils/camelToSnakeCase.ts
@@ -1,7 +1,7 @@
-export const camelToSnakeCaseKey = (key: string) => 
+export const camelToSnakeCaseKey = (key: string): string => 
   key.replace(/[A-Z]/g, upperCaseLetter => `_${upperCaseLetter.toLowerCase()}`);
 
-export const camelToSnakeCaseObject = (obj: any) => { 
+export const camelToSnakeCaseObject = (obj: any): any => { 
   const snakeCasedObj = {};
   for (const key in obj) {
       if (obj.hasOwnProperty(key)) {
